Desktop/Coding/limitless-sync-HR-AR/requirements.txt
requests==2.31.0
GitPython==3.1.40
schedule==1.2.0
python-dotenv==1.0.0
